+++ Resources/CreateDistPackage.sh
@@ -3,8 +3,8 @@
 # Orthanc - A Lightweight, RESTful DICOM Store
 # Copyright (C) 2012-2016 Sebastien Jodogne, Medical Physics
 # Department, University Hospital of Liege, Belgium
-# Copyright (C) 2017-2022 Osimis S.A., Belgium
-# Copyright (C) 2021-2022 Sebastien Jodogne, ICTEAM UCLouvain, Belgium
+# Copyright (C) 2017-2024 Osimis S.A., Belgium
+# Copyright (C) 2021-2024 Sebastien Jodogne, ICTEAM UCLouvain, Belgium
 #
 # This program is free software: you can redistribute it and/or
 # modify it under the terms of the GNU Affero General Public License
@@ -52,4 +52,4 @@ docker run -t ${DOCKER_FLAGS} --rm \
        -v ${ROOT_DIR}/WebApplication:/source/WebApplication:ro \
        -v ${ROOT_DIR}/WebApplication/dist/:/target:rw \
        ${IMAGE} \
-       bash /source/build.sh ${VERSION}
+       bash /source/build.sh