+++ WebApplication/src/components/InstanceDetails.vue
@@ -37,7 +37,7 @@ export default {
         <tr>
             <td width="80%" class="cut-text">
                 <TagsTree 
-                    :tags="tags">
+                    :tags="headers">
                 </TagsTree>
             </td>
             <td width="20%" class="instance-button-group">
@@ -48,8 +48,16 @@ export default {
                 :seriesMainDicomTags="this.seriesMainDicomTags"
                 :patientMainDicomTags="this.patientMainDicomTags"
                 :instanceTags="this.tags"
+                :instanceHeaders="this.headers"
                 @deletedResource="onDeletedInstance"
                 ></ResourceButtonGroup>
+            </td>
+        </tr>
+        <tr>
+            <td width="80%" class="cut-text">
+                <TagsTree 
+                    :tags="tags">
+                </TagsTree>
             </td>
         </tr>
     </table>
@@ -60,13 +68,22 @@ export default {
     margin-top: var(--details-top-margin);
     margin-left: 5%;
     width: 95% !important;
-    background-color: var(--instance-selected-color);
+    background-color: var(--instance-details-bg-color) !important;
     font-size: 0.8rem;
 }
 
 .instance-details-table td {
     vertical-align: top;
 }
+
+.instance-details-table>:not(caption) >* >* {
+  background-color: var(--instance-details-bg-color) !important;
+}
+
+.instance-details-table >* >* {
+  background-color: var(--instance-details-bg-color) !important;
+}
+
 
 .details-label {
     font-weight: 700;