+++ WebApplication/vite.config.js
@@ -18,7 +18,7 @@ export default defineConfig({
         main: resolve(__dirname, 'index.html'),
         landing: resolve(__dirname, 'token-landing.html'),
         retrieve: resolve(__dirname, 'retrieve-and-view.html')
-      }
+      },
     }
   },
   css: {